+++ tests/unit/test_semver.nim
@@ -8,6 +8,7 @@ proc main() =
 
   doAssert parseVersion("0.1") == parseVersion("0.1.0")
   doAssert parseVersion("0.1.0") == parseVersion("0.1.0")
+  doAssert parseVersion("0.1.05") == parseVersion("0.1.5")
   doAssert not (parseVersion("0.1") == parseVersion("0.1.5"))
   doAssert not (parseVersion("0.1") == parseVersion("0.1-dev"))
   doAssert not (parseVersion("0.1.0") == parseVersion("0.1.5"))
@@ -33,12 +34,14 @@ proc main() =
   doAssert parseVersion("0.1") > parseVersion("0.1-dev")
   doAssert parseVersion("0.1.5") > parseVersion("0.1")
   doAssert parseVersion("0.1.5") > parseVersion("0.1.0")
+  doAssert parseVersion("0.1.05") > parseVersion("0.1.0")
   doAssert not (parseVersion("0.1") > parseVersion("0.1"))
   doAssert not (parseVersion("0.1.0") > parseVersion("0.1"))
 
   doAssert parseVersion("0.1") >= parseVersion("0.1-dev")
   doAssert parseVersion("0.1.5") >= parseVersion("0.1")
   doAssert parseVersion("0.1.5") >= parseVersion("0.1.0")
+  doAssert parseVersion("0.1.05") >= parseVersion("0.1.0")
   doAssert parseVersion("0.1") >= parseVersion("0.1")
   doAssert parseVersion("0.1.0") >= parseVersion("0.1")